+++ src/lib-public/generators/FA2/Wiersze.ts
@@ -112,10 +112,23 @@ export function generateWiersze(faVat: Fa): Content {
       alignment: Position.RIGHT,
       margin: [0, 8, 0, 0],
     };
-  } else if ((rodzajFaktury == TRodzajFaktury.KOR || rodzajFaktury == TRodzajFaktury.KOR_ROZ) && hasP15) {
+  } else if (rodzajFaktury == TRodzajFaktury.KOR && hasP15) {
     opis = {
       stack: createLabelTextArray([
         { value: i18n.t('invoice.rows.totalAmountDueCorrection'), formatTyp: FormatTyp.LabelGreater },
+        {
+          value: p_15,
+          formatTyp: [FormatTyp.CurrencyGreater, FormatTyp.HeaderContent, FormatTyp.Value],
+          currency: getValue(faVat.KodWaluty)?.toString() ?? '',
+        },
+      ]),
+      alignment: Position.RIGHT,
+      margin: [0, 8, 0, 0],
+    };
+  } else if (rodzajFaktury == TRodzajFaktury.KOR_ROZ && hasP15) {
+    opis = {
+      stack: createLabelTextArray([
+        { value: i18n.t('invoice.rows.remainingAmountCorrection'), formatTyp: FormatTyp.LabelGreater },
         {
           value: p_15,
           formatTyp: [FormatTyp.CurrencyGreater, FormatTyp.HeaderContent, FormatTyp.Value],